+++ GadgetEngine/Data/List.h
@@ -74,7 +74,7 @@ namespace Gadget{
 		inline Node* Add(const T& value_, Node* nodePos_){
 			GADGET_BASIC_ASSERT(IsEmpty() || nodePos_ != nullptr);
 			GADGET_BASIC_ASSERT(nodePos_ == nullptr || ContainsNode(nodePos_));
-			GADGET_BASIC_ASSERT(size < std::numeric_limits<size_t>::max());
+			GADGET_BASIC_ASSERT(size < std::numeric_limits<int64_t>::max());
 
 			Node* newNode = new Node(value_);
 
@@ -105,7 +105,7 @@ namespace Gadget{
 		}
 
 		inline Node* AddFront(const T& value_){
-			GADGET_BASIC_ASSERT(size < std::numeric_limits<size_t>::max());
+			GADGET_BASIC_ASSERT(size < std::numeric_limits<int64_t>::max());
 
 			Node* newNode = new Node(value_);
 
@@ -239,12 +239,12 @@ namespace Gadget{
 			head = prev;
 		}
 
-		constexpr inline size_t Size() const{ return size; }
+		constexpr inline int64_t Size() const{ return size; }
 		constexpr inline bool IsEmpty() const{ return size == 0; }
 		constexpr inline Node* Front() const{ return head; }
 		constexpr inline Node* Back() const{ return tail; }
 
-		constexpr inline Node* GetNode(size_t index_) const{
+		constexpr inline Node* GetNode(int64_t index_) const{
 			GADGET_BASIC_ASSERT(index_ < size);
 
 			if(index_ == 0){
@@ -256,7 +256,7 @@ namespace Gadget{
 			}
 
 			Node* current = head;
-			for(size_t i = 1; i <= index_; i++){
+			for(int64_t i = 1; i <= index_; i++){
 				current = current->next;
 			}
 
@@ -264,7 +264,7 @@ namespace Gadget{
 			return current;
 		}
 
-		constexpr inline Node* GetNthLastNode(size_t index_) const{
+		constexpr inline Node* GetNthLastNode(int64_t index_) const{
 			GADGET_BASIC_ASSERT(index_ <= size);
 			GADGET_BASIC_ASSERT(index_ > 0);
 			if(index_ > size || index_ == 0){
@@ -277,7 +277,7 @@ namespace Gadget{
 			}
 
 			Node* current = nullptr;
-			for(size_t i = 0; i < size + 1 - index_; i++){
+			for(int64_t i = 0; i < size + 1 - index_; i++){
 				if(current == nullptr){
 					current = head;
 				}else{
@@ -324,7 +324,7 @@ namespace Gadget{
 			}
 
 			Node* current = head;
-			size_t count = 0;
+			int64_t count = 0;
 			while(current != nullptr){
 				count++;
 				if(count > Size()){
@@ -348,7 +348,7 @@ namespace Gadget{
 		constexpr const Iterator end() const{ return Iterator(nullptr); }
 
 	private:
-		size_t size;
+		int64_t size;
 		Node* head;
 		Node* tail;
