+++ GadgetEngine/Data/String.h
@@ -7,7 +7,7 @@
 namespace Gadget{
 	class String{
 	public:
-		String(const char* str_ = "") : data(Math::Clamp<size_t>(16, std::numeric_limits<size_t>::max(), std::strlen(str_))){
+		String(const char* str_ = "") : data(Math::Clamp<int64_t>(16, std::numeric_limits<int64_t>::max(), std::strlen(str_))){
 			data.Add('\0');
 			Append(str_);
 		}
@@ -17,7 +17,7 @@ namespace Gadget{
 			Append(strArray_);
 		}
 
-		explicit String(size_t capacity_) : data(capacity_ + 1){
+		explicit String(int64_t capacity_) : data(capacity_ + 1){
 			data.Add('\0');
 		}
 
@@ -63,14 +63,14 @@ namespace Gadget{
 		void Append(const Array<String>& strs_){
 			//This technically has higher time complexity than just doing one loop,
 			//but doing only one memory allocation is worth it if we're appending lots of strings/large strings
-			size_t newLength = Length();
-			for(size_t i = 0; i < strs_.Size(); i++){
+			int64_t newLength = Length();
+			for(int64_t i = 0; i < strs_.Size(); i++){
 				newLength += strs_[i].Length();
 			}
 
 			data.Reserve(newLength);
 
-			for(size_t i = 0; i < strs_.Size(); i++){
+			for(int64_t i = 0; i < strs_.Size(); i++){
 				Append(strs_[i].Value());
 			}
 		}
@@ -97,7 +97,7 @@ namespace Gadget{
 		void operator+=(char c_){ *this = *this + c_; }
 		void operator+=(int32_t number_){ *this = *this + number_; }
 
-		void InsertAt(size_t index_, char c){
+		void InsertAt(int64_t index_, char c){
 			if(index_ >= Length()){
 				Append(c);
 				return;
@@ -106,7 +106,7 @@ namespace Gadget{
 			data.InsertAt(index_, c);
 		}
 
-		constexpr void RemoveAt(size_t index_){
+		constexpr void RemoveAt(int64_t index_){
 			if(index_ >= Length()){
 				return;
 			}
@@ -124,7 +124,7 @@ namespace Gadget{
 			data.RemoveAll(value_);
 		}
 
-		void Reserve(size_t capacity){
+		void Reserve(int64_t capacity){
 			data.Reserve(capacity + 1); //Make sure there's room for the null terminator!
 		}
 
@@ -137,7 +137,7 @@ namespace Gadget{
 		}
 		
 		void Trim(){
-			size_t whitespaceAtStart;
+			int64_t whitespaceAtStart;
 			for(whitespaceAtStart = 0; whitespaceAtStart < Length(); whitespaceAtStart++){
 				if(!IsWhitespace(data[whitespaceAtStart])){
 					break;
@@ -149,7 +149,7 @@ namespace Gadget{
 			GADGET_BASIC_ASSERT(data.Size() >= 1);
 
 			if(Length() > 0){
-				GADGET_BASIC_ASSERT(data.Size() < static_cast<uint64_t>(std::numeric_limits<int64_t>::max())); //I doubt this will ever be an issue, but just in case
+				GADGET_BASIC_ASSERT(data.Size() < std::numeric_limits<int64_t>::max()); //I doubt this will ever be an issue, but just in case
 				int64_t indexToEndWhitespace;
 				for(indexToEndWhitespace = data.Size() - 2; indexToEndWhitespace >= 0; indexToEndWhitespace--){
 					if(!IsWhitespace(data[indexToEndWhitespace])){
@@ -165,11 +165,11 @@ namespace Gadget{
 		void FindAndReplace(char find_, const String& replace_){
 			GADGET_BASIC_ASSERT(find_ != '\0'); //Do not try to replace the null terminator
 
-			for(size_t i = 0; i < Length(); i++){
+			for(int64_t i = 0; i < Length(); i++){
 				if(data[i] == find_){
 					RemoveAt(i);
 
-					for(size_t j = 0; j < replace_.Length(); j++){
+					for(int64_t j = 0; j < replace_.Length(); j++){
 						InsertAt(i + j, replace_[j]);
 					}
 				}
@@ -177,18 +177,18 @@ namespace Gadget{
 		}
 
 		constexpr void ToLower(){
-			for(size_t i = 0; i < Length(); i++){
+			for(int64_t i = 0; i < Length(); i++){
 				data[i] = static_cast<char>(std::tolower(data[i]));
 			}
 		}
 
 		constexpr void ToUpper(){
-			for(size_t i = 0; i < Length(); i++){
+			for(int64_t i = 0; i < Length(); i++){
 				data[i] = static_cast<char>(std::toupper(data[i]));
 			}
 		}
 
-		constexpr inline int64_t Find(char c_, size_t startPos_ = 0) const{ return data.Find(c_, startPos_); }
+		constexpr inline int64_t Find(char c_, int64_t startPos_ = 0) const{ return data.Find(c_, startPos_); }
 
 		constexpr inline bool Contains(char c) const{ return data.Contains(c); }
 
@@ -198,11 +198,11 @@ namespace Gadget{
 				return false;
 			}
 
-			for(size_t i = 0; i < data.Size(); i++){
+			for(int64_t i = 0; i < data.Size(); i++){
 				bool isEqual = true;
 
-				size_t j = i;
-				for(size_t k = 0; k < str_.Length(); j++, k++){
+				int64_t j = i;
+				for(int64_t k = 0; k < str_.Length(); j++, k++){
 					if(data[j] != str_[k]){
 						GADGET_ASSERT(str_[k] != '\0', "String comparison failing due to null terminator!");
 
@@ -219,7 +219,7 @@ namespace Gadget{
 			return false;
 		}
 
-		String SubString(size_t startIndex_, size_t endIndex_) const{
+		String SubString(int64_t startIndex_, int64_t endIndex_) const{
 			GADGET_BASIC_ASSERT(startIndex_ != endIndex_);
 			return String(data.SubRange(startIndex_, endIndex_));
 		}
@@ -241,11 +241,11 @@ namespace Gadget{
 		}
 
 		bool operator==(const char* str_) const{
-			if(Length() != std::strlen(str_)){
+			if(Length() != static_cast<int64_t>(std::strlen(str_))){
 				return false;
 			}
 
-			for(size_t i = 0; i < Length(); i++){
+			for(int64_t i = 0; i < Length(); i++){
 				if(data[i] != str_[i]){
 					return false;
 				}
@@ -258,17 +258,17 @@ namespace Gadget{
 			return data == str_.data;
 		}
 
-		constexpr const char& operator[](size_t i_) const{
+		constexpr const char& operator[](int64_t i_) const{
 			GADGET_BASIC_ASSERT(i_ < data.Size());
 			return data[i_];
 		}
 
-		constexpr char& operator[](size_t i_){
+		constexpr char& operator[](int64_t i_){
 			GADGET_BASIC_ASSERT(i_ < data.Size());
 			return data[i_];
 		}
 
-		constexpr size_t Length() const{ return data.Size() - 1; }
+		constexpr int64_t Length() const{ return data.Size() - 1; }
 
 		constexpr bool IsEmpty() const{
 			GADGET_BASIC_ASSERT(data.Size() > 0);
